input: focaltech_touch_mi: Enable support for pressure
Touchscreen firmware used in Pyxis reports pressure values for
touches so provide them up to userspace and remove relevant config
knob as well.
Also do not report the pressure for FOD event because the firmware
does not provide pressure value in this case and currently reports
value of touching area size as pressure value.

Signed-off-by: Ivan Vecera <ivan@cera.cz>
Change-Id: I6c748ec74c674ec8dafe1f68f7803ec20d90eb9c